refactor: migrate e2e test
ref #2318
